Set option to enable express server
Add the config electron viewer file
Refactoring
Set new menu button

Button to return to home page when user click on external link.
Set menu button to return on previous page
