+++ package.json
@@ -36,6 +36,7 @@
     "electron-builder": "^22.14.5",
     "electron-packager": "^15.4.0",
     "filemanager-webpack-plugin": "^6.1.7",
+    "nodemon": "^2.0.19",
     "save-remote-file-webpack-plugin": "^1.1.0",
     "webpack-merge": "^5.8.0",
     "webpack-nano": "^1.1.1",
@@ -50,7 +51,6 @@
     "express": "^4.17.3",
     "express-json-validator-middleware": "^2.2.1",
     "lunr": "^2.3.9",
-    "nodemon": "^2.0.16",
     "path-browserify": "^1.0.1",
     "replace-in-file-webpack-plugin": "^1.0.6",
     "rimraf": "^3.0.2",
