+++ src/server.js
@@ -8,7 +8,7 @@ const path = require('path');
 const app = express();
 
 const {getLogTransportConsole, getLogTransportFile, getLogResolvePath} = require('./lib/logFormat');
-const ConstructURL = require('./utils/ConstructURL');
+const ConfigElectronViewer = require('./utils/ConfigElectronViewer');
 const ContentIndexer = require('./utils/lunr/ContentIndexer');
 const SearchHandler = require('./utils/lunr/SearchHandler');
 const { validate } = new Validator();
@@ -26,7 +26,9 @@ app
 .use(bodyParser.json())
 .use(express.static(path.join(__dirname, '../static')));
 
+// Index the content for lunr
 ContentIndexer.tocIndexer();
+// Create index from the documentation
 ContentIndexer.createIndex();
 
 log.debug("Product ID = [" + ContentIndexer.getMultiDocContent().productId + "]; Plugin ID = [" + ContentIndexer.getMultiDocContent().pluginId + "];")
@@ -41,7 +43,11 @@ const urlSchema = {
     },
 };
 
+// POST Request sended by the help viewer to redirect to the clicked link from the popover
 app.post('/viewerurl', validate({body: urlSchema}),(req, res) => {
+    let splitUrl = req.body.url.split('/');
+    let languageCode = splitUrl[splitUrl.length -1];
+
     res.setHeader('Content-Type', 'application/json');
 
     if(res.status === 500){
@@ -54,9 +60,6 @@ app.post('/viewerurl', validate({body: urlSchema}),(req, res) => {
 
     ipc.send('requested-url', req.body.url);
 
-    let splitUrl = req.body.url.split('/');
-    let languageCode = splitUrl[splitUrl.length -1];
-
     if(languageCode == "fr"){
         message = 'Chargement ...';
     }
@@ -65,13 +68,19 @@ app.post('/viewerurl', validate({body: urlSchema}),(req, res) => {
     res.send(`POST request body ${JSON.stringify({url: req.body.url})} was sending succesfully`);
 });
 
+// GET Request to find the indexed content from lunr
 app.get('/httpd/api/search', (req, res) => {
     let query = req.query['query'];
     let exactMatch = req.query['exact-match'];
     let lang = req.query['lang'];
     let limit = parseInt(req.query['limit']);
-
+    let getSearchResults = SearchHandler.getSearchResults(query, exactMatch, lang ? lang : "en");
     res.setHeader('Content-Type', 'application/json');
+
+    if(limit && limit > 0){
+        getSearchResults = getSearchResults.slice(Math.max(getSearchResults.length - limit, 0));
+    }
+
     if(res.status === 500){
         log.error(req, res);
         res.send('Internal Server Error');
@@ -80,12 +89,6 @@ app.get('/httpd/api/search', (req, res) => {
         res.status(404).json({message});
     }
     
-    let getSearchResults = SearchHandler.getSearchResults(query, exactMatch, lang ? lang : "en");
-    
-    if(limit && limit > 0){
-        getSearchResults = getSearchResults.slice(Math.max(getSearchResults.length - limit, 0));
-    }
-
     log.debug("Query parameters: Query=["+ query +"], exactMatch=["+ exactMatch +"], lang=["+ lang +"], limit=["+ limit +"]");
     log.debug("Result length = " + getSearchResults.length);
 
@@ -97,6 +100,6 @@ app.use(({res}) => {
     res.send(message);
 });
 
-var server = app.listen(ConstructURL.getServerPort(), () => {
-    log.info(`Server listening on port : ${ConstructURL.getServerPort()}`);
+var server = app.listen(ConfigElectronViewer.getServerPort(), () => {
+    log.info(`Server listening on port : ${ConfigElectronViewer.getServerPort()}`);
 });
