+++ src/utils/lunr/SearchHandler.js
@@ -38,5 +38,4 @@ class SearchHandler {
         });
     }
 }
-
 module.exports = SearchHandler;
